+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## 2.0.0 - 2026-07-03
+- Major update: support for multiple companies added
+- Updated login screen with new app logo & color scheme
+
+
 ## 1.2.6 - 2026-06-27
 - Fixed bug where completion email was not being sent to admin team
 - Fixed bug where on mobile UI, exp date field was too large
+
 
 ## 1.2.5 - 2026-06-26
 - Added gender marker to employee personal information